+++ .storybook/theme.ts
@@ -1,0 +1,13 @@
+import { create } from '@storybook/theming';
+import { light } from '@sumup/design-tokens';
+
+export const theme = create({
+  base: 'light',
+  brandTitle: 'Circuit UI',
+  brandUrl: 'https://github.com/sumup-oss/circuit-ui',
+  brandImage: '/images/logo-name.svg',
+  fontBase: light.fontStack.default,
+  colorPrimary: light.colors.p500,
+  colorSecondary: light.colors.p500,
+  appBg: light.colors.bodyBg,
+});
