+++ docs/introduction/1-welcome.mdx
@@ -23,7 +23,7 @@ import { Meta, Image, Intro, Teaser } from '../../.storybook/components';
 
 Set up a new app with Circuit UI or add it to an existing project.
 
-[Get started](Introduction/Getting-Started/Page)
+[Get started →](Introduction/Getting-Started/Page)
 
 </Teaser>
 
@@ -31,7 +31,7 @@ Set up a new app with Circuit UI or add it to an existing project.
 
 File a bug report, suggest a change, or open a pull request.
 
-[Contribute](Contributing/Overview)
+[Contribute →](Contributing/Overview)
 
 </Teaser>
 
@@ -39,7 +39,7 @@ File a bug report, suggest a change, or open a pull request.
 
 Learn about our foundations such as colors, spacing, and typography.
 
-[Browse theme reference](Features/Theme/Page)
+[Browse theme reference →](Features/Theme/Page)
 
 </Teaser>
 
@@ -47,11 +47,11 @@ Learn about our foundations such as colors, spacing, and typography.
 
 Explore available icons in the Circuit UI icons library.
 
-[Browse icons](Features/Icons/Page)
+[Browse icons →](Features/Icons/Page)
 
 </Teaser>
 
-<div style={{ clear: 'both' }} />
+<div style={{ clear: 'both', height: '2rem' }} />
 
 ## Code of conduct
